% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinOutModel_class.R
\docType{data}
\name{Abstract_BinDat}
\alias{Abstract_BinDat}
\title{(NOT USED) Abstract summary measure class for P(sA[j]|sW,sA[j])}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fa7c4c53d10> 
 - attr(*, "name")= chr "Abstract_BinDat_generator"
}}
\usage{
Abstract_BinDat
}
\description{
(NOT USED) Abstract summary measure class for P(sA[j]|sW,sA[j])
}
\keyword{datasets}

