% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinOutModel_class.R
\docType{data}
\name{BinOutModel}
\alias{BinOutModel}
\title{Class for fitting a logistic model with binary outcome, P(sA[j]|sW,sA[j])}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fd7f9f7a830> 
 - attr(*, "name")= chr "BinOutModel_generator"
}}
\usage{
BinOutModel
}
\description{
Class for fitting a logistic model with binary outcome, P(sA[j]|sW,sA[j])
}
\keyword{datasets}

