% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DatNet_class.R
\docType{class}
\name{DatNet}
\alias{DatNet}
\title{Class for storing netW (or netA) (network variables for all W (or A) variables)}
\format{An R6 class object.}
\usage{
DatNet
}
\description{
Class for storing netW (or netA) (network variables for all W (or A) variables)
}
\details{
Following fields are created during initialization
\itemize{
\item{Kmax} ...
\item{netind_cl} ...
\item{subset_regs} ...
\item{sA_nms} ...
\item{sW_nms} ...
\item{Kmax} ...
}
Class for evaluating and storing arbitrary summary measures sVar.
The summary measures are evaluated based on the user-specified sVar expressions in sVar.object (sW or sA),
in the environment of the input data.frame (Odata).
The evaluated summary measures from sVar.object are stored as a matrix (self$mat.sVar).
Contains methods for replacing missing values with default in gvars$misXreplace.
Also contains method for detecting /setting sVar variable type (binary, categor, contin).
For continous sVar this class provides methods for detecting / setting bin intervals, normalization, disretization and construction of bin indicators.
}
\keyword{datasets}

