% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DatNet_class.R
\docType{class}
\name{DatNet.sWsA}
\alias{DatNet.sWsA}
\title{Class for managing combined summary measures sW & sA from DatNet class.}
\format{An R6 class object.}
\usage{
DatNet.sWsA
}
\description{
Class for managing combined summary measures sW & sA from DatNet class.
}
\details{
Following fields are created during initialization
\itemize{
\item{nodes} ...
\item{datnetW} ...
\item{datnetA} ...
\item{YnodeVals} ...
\item{det.Y} ...
\item{Kmax} ...
}
Methods for combining, subsetting, discretizing & binirizing of summary measures in sW & sA. Inherits from DatNet class.
The combined dataset of all (sW, sA) summary measures returned by DatNet is stored as a matrix under self$mat.sVar
}
\keyword{datasets}

