% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmlenet-package.R
\docType{data}
\name{datO_input}
\alias{datO_input}
\title{An example simulated dataset used for testing}
\usage{
data(datO_input)
}
\description{
An example simulated dataset used for testing
}
\keyword{datasets}

