% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DefineSummariesClass.R
\docType{class}
\name{DefineSummariesClass}
\alias{DefineSummariesClass}
\title{R6 class for parsing and evaluating user-specified summary measures (in \code{exprs_list})}
\format{An \code{\link{R6Class}} generator object}
\usage{
DefineSummariesClass
}
\description{
This R6 class can parse and evaluate (given the input data) the summary measures defined with functions
 \code{\link{def.sW}} and \code{\link{def.sA}}.
 The summary measure expressions (stored in \code{exprs_list}) are evaluated in the environment of the input data.frame.
 Note that the resulting summary measures are never stored inside this class,
 the data is stored only in \code{\link{DatNet}} and \code{\link{DatNet.sWsA}} R6 classes.
}
\details{
\itemize{
\item{\code{type}} - Type of the summary measure, sW or sA, determined by the calling functions \code{def.sW} or \code{def.sA}.
\item{\code{exprs_list}} - Deparsed list of summary measure expressions (as strings).
\item{\code{new_expr_names}} - Re-evaluated summary measure names, if none were provided by the user these will be
 evaluated on the basis of variable names used in the expression.
\item{\code{sVar.names.map}} - Map of the summary measure names given by the user to actual column names in the summary measure matrix.
}
}
\section{Methods}{

\describe{
  \item{\code{new(type)}}{...}
  \item{\code{set.new.exprs(exprs_list)}}{...}
  \item{\code{add.new.exprs(NewSummaries)}}{...}
  \item{\code{eval.nodeforms(data.df, netind_cl)}}{...}
  \item{\code{df.names(data.df)}}{List of variables in the input data \code{data.df} gets assigned to a special
  variable (\code{ANCHOR_ALLVARNMS_VECTOR_0})}
}
}
\keyword{R6}
\keyword{class}

