% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Model_h.R
\docType{function}
\name{fit.hbars}
\alias{fit.hbars}
\title{Defining and fitting the clever covariate h under g_0 and g_star, i.e. models P(sA[j] | sW,sA[j])}
\usage{
fit.hbars(datNetObs, est_params_list)
}
\description{
Defining and fitting the clever covariate h under g_0 and g_star, i.e. models P(sA[j] | sW,sA[j])
}

