% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmAlgorithmClass.R
\docType{class}
\name{logisfitR6}
\alias{logisfitR6}
\title{logisfitR6}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
logisfitR6
}
\description{
logisfitR6
}
\keyword{datasets}
