% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcEvalPsi_class.R
\docType{class}
\name{mcEvalPsi}
\alias{mcEvalPsi}
\title{Class for Monte-Carlo evaluation of various substitution estimators under user-specified stochastic intervention gstar.}
\format{An R6 class object.}
\usage{
mcEvalPsi
}
\description{
Class for Monte-Carlo evaluation of various substitution estimators under user-specified stochastic intervention gstar.
}
\details{
Following fields are created during initialization
\itemize{
\item{nodes} ...
\item{subset_regs} ...
\item{sA_nms} ...
\item{sW_nms} ...
\item{Kmax} ...
}
Class for evaluating and storing arbitrary summary measures sVar.
}
\keyword{datasets}

