% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DatNetClass.R
\docType{class}
\name{OdataDT}
\alias{OdataDT}
\title{R6 class for storing and managing the user-input data.}
\format{An \code{\link{R6Class}} generator object}
\usage{
OdataDT
}
\description{
The data is stored as a data.table under the public field \code{OdataDT}.
}
\details{
\itemize{
\item{\code{nOdata}} - Maximum number of friends for any observation.
\item{\code{nOdata}} - number of observations in the observed data frame.
\item{\code{nFnode}} - Name of the vector that stores the number of friends for each observation (always set to 'nF').
\code{iid_data_flag} - Is the input data iid (i.e., no network)?
\item{\code{Wnodes}} - Names of the baseline covariates (not used).
\item{\code{Anodes}} - Names of the intervention nodes.
\item{\code{Lnodes}} - Names of the time-varying covariates (not used).
\item{\code{Ynodes}} - Names of the outcome nodes (not used).
\item{\code{A.g0.mat}} - Matrix of observed exposures
\item{\code{A.gstar.mat}} - Matrix of counterfactual exposures.
}
}
\section{Methods}{

\describe{
  \item{\code{new(Odata, nFnode, iid_data_flag, nodes, ...)}}{...}
  \item{\code{replaceVar(varName, newVarVal)}}{...}
  \item{\code{replaceOneAnode(AnodeName, newAnodeVal)}}{...}
  \item{\code{replaceManyAnodes(Anodes, newAnodesMat)}}{...}
}
}

\section{Active Bindings}{

\describe{
   \item{\code{ncols.sVar}}{...}
   \item{\code{emptydat.sVar}}{...}
   \item{\code{nodes}}{...}
   \item{\code{dat.sVar}}{...}
   \item{\code{emptydat.sVar}}{...}
   \item{\code{nodes}}{...}
}
}
\keyword{R6}
\keyword{class}

