% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Model_h.R
\name{pred.hbars}
\alias{pred.hbars}
\title{Predict the h weights under g_0 and g_star using existing m.h.fit model fit}
\usage{
pred.hbars(newdatnet = NULL, m.h.fit)
}
\description{
Predict the h weights under g_0 and g_star using existing m.h.fit model fit
}

