% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{print.tmlenet.opts}
\alias{print.tmlenet.opts}
\title{Print Current Option Settings for \code{tmlenet}}
\usage{
\method{print}{tmlenet.opts}()
}
\value{
Invisibly returns a list of \code{tmlenet} options.
}
\description{
Print Current Option Settings for \code{tmlenet}
}
\seealso{
\code{\link{tmlenet.options}}
}

