% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_tmlenet_opts}
\alias{print_tmlenet_opts}
\title{Print Current Option Settings for \code{tmlenet}}
\usage{
print_tmlenet_opts()
}
\value{
Invisibly returns a list of \code{tmlenet} options.
}
\description{
Print Current Option Settings for \code{tmlenet}
}
\seealso{
\code{\link{tmlenet_options}}
}

