% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmAlgorithmClass.R
\docType{class}
\name{speedglmR6}
\alias{speedglmR6}
\title{speedglmR6}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
speedglmR6
}
\description{
speedglmR6
}
\keyword{datasets}
