% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummariesModel_class.R
\docType{class}
\name{SummariesModel}
\alias{SummariesModel}
\title{R6 class for fitting and predicting model P(sA|sW) under g.star or g.0}
\format{An R6 class object.}
\usage{
SummariesModel
}
\description{
R6 class for fitting and predicting model P(sA|sW) under g.star or g.0
}
\details{
Class for defining, fitting and predicting the probability model P(sA = sa | sW = sw) under g_star or under g_0 for summary measures (sW,sA).
\itemize{
\item{n_regs} ...
\item{sA_nms} ...
\item{sW_nms} ...
}
Additional details about implementation of the class...
}
\keyword{datasets}

