% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{tmlenet.options}
\alias{tmlenet.options}
\title{Setting Options for \code{tmlenet}}
\usage{
tmlenet.options(useglm = FALSE, nbins = 15, maxncats = 5,
  binByMass = TRUE, poolContinVar = FALSE, maxNperBin = 1000)
}
\arguments{
\item{useglm}{Set to \code{FALSE} to estimate with \code{speedglm::speedglm.wfit} and \code{TRUE} for \code{glm::glm.fit}.}

\item{nbins}{Set the default number of bins to use when binning a continous outcome variable.}

\item{maxncats}{Max number of unique categories a categorical variable \code{sA[j]} can have.
If \code{sA[j]} has more it is automatically considered continuous.}

\item{binByMass}{Define bin cutoffs for a continuous outcome using even mass distribution?
Setting to \code{TRUE} allows for data-adaptive bin cutoff definitions,
where each bin will be defined so that it contains an approximately the same number of observations across all bins.}

\item{poolContinVar}{Set to \code{TRUE} for fitting a pooled regression which pools bin indicators across all bins.
When fitting a model for binirized continuous outcome, set to \code{TRUE}
for pooling bin indicators across several bins into one outcome regression?}

\item{maxNperBin}{Max number of observations per 1 bin for a continuous outcome (only applies when \code{binByMass=TRUE})}
}
\value{
Invisibly returns a list with old option settings.
}
\description{
Additional options that control the estimation algorithm in \code{tmlenet} package
}
\seealso{
\code{\link{print.tmlenet.opts}}
}

