% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinOutModel_class.R
\docType{class}
\name{BinDat}
\alias{BinDat}
\title{#}
\format{An \code{\link{R6Class}} generator object}
\usage{
BinDat
}
\description{
#
}
\details{
\itemize{
\item{bin_names} - Matrix of friend indices (network IDs) of \code{dim = (nobs x Kmax)}.
\item{ID} - Vector of integers, where \code{nF[i]} is the integer number of friends (0 to \code{Kmax}) for observation \code{i}.
\item{pooled_bin_name} - Number of observations
\item{nbins} - Maximum number of friends for any observation.
\item{outvar} - Maximum number of friends for any observation.
\item{predvars} - Maximum number of friends for any observation.
\item{pool_cont} - Maximum number of friends for any observation.
\item{outvars_to_pool} - Maximum number of friends for any observation.
\item{subset_expr} - Maximum number of friends for any observation.
\item{subset_idx} - Maximum number of friends for any observation.
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg)}}{Uses \code{reg} R6 \code{\link{RegressionClass}} object to instantiate a new storage container for a
  design matrix and binary outcome.}
  \item{\code{show()}}{ Print information on outcome and predictor names used in this regression model}
  \item{\code{newdata()}}{...}
  \item{\code{define.subset_idx(...)}}{...}
  \item{\code{setdata()}}{...}
  \item{\code{logispredict()}}{...}
  \item{\code{setdata.long()}}{...}
  \item{\code{logispredict.long()}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{emptydata}}{...}
  \item{\code{emptyY}}{...}
  \item{\code{emptySubset_idx}}{...}
  \item{\code{emptyN}}{...}
  \item{\code{getXmat}}{...}
  \item{\code{getY}}{...}
}
}
\keyword{&}
\keyword{(fitting}
\keyword{Bin}
\keyword{Can}
\keyword{DatNet.sWsA}
\keyword{P(Bin|Xmat).}
\keyword{R6}
\keyword{The}
\keyword{This}
\keyword{Xmat,}
\keyword{a}
\keyword{across}
\keyword{also}
\keyword{and}
\keyword{appropriate}
\keyword{as}
\keyword{be}
\keyword{binary}
\keyword{can}
\keyword{class}
\keyword{columns}
\keyword{converting}
\keyword{data}
\keyword{design}
\keyword{e.g.,}
\keyword{for}
\keyword{format}
\keyword{get}
\keyword{has}
\keyword{in}
\keyword{indicators}
\keyword{indicators)}
\keyword{internal}
\keyword{logistic}
\keyword{long}
\keyword{manage}
\keyword{matrix}
\keyword{methods}
\keyword{model}
\keyword{one}
\keyword{outcome}
\keyword{perform}
\keyword{pooled}
\keyword{pooling}
\keyword{queries}
\keyword{regression}
\keyword{request,}
\keyword{requested,}
\keyword{row}
\keyword{several}
\keyword{single}
\keyword{storage}
\keyword{store}
\keyword{storing}
\keyword{subsets}
\keyword{that}
\keyword{the}
\keyword{to}
\keyword{used}
\keyword{well}
\keyword{when}
\keyword{wide}

