% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinOutModel_class.R
\docType{data}
\name{BinDat}
\alias{BinDat}
\title{#}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fe0a8439838> 
 - attr(*, "name")= chr "BinDat_generator"
}}
\usage{
BinDat
}
\description{
#
}
\keyword{&}
\keyword{(when}
\keyword{/}
\keyword{Converts}
\keyword{DatNet.sWsA}
\keyword{Data}
\keyword{P(sA[j]}
\keyword{Queries}
\keyword{across}
\keyword{appropriate}
\keyword{binary}
\keyword{class}
\keyword{columns}
\keyword{data}
\keyword{fitting}
\keyword{for}
\keyword{get}
\keyword{indicators}
\keyword{indicators)}
\keyword{internal}
\keyword{long}
\keyword{manipulation}
\keyword{measures}
\keyword{model}
\keyword{modeling}
\keyword{one}
\keyword{pooling}
\keyword{row}
\keyword{sW,bar{sA}[j-1]).}
\keyword{several}
\keyword{storage}
\keyword{subsets}
\keyword{summary}
\keyword{to}
\keyword{wide}
\keyword{|}

