% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinOutModel_class.R
\docType{data}
\name{BinDat}
\alias{BinDat}
\title{#}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fe745ad0ca8> 
 - attr(*, "name")= chr "BinDat_generator"
}}
\usage{
BinDat
}
\description{
#
}
\keyword{(as}
\keyword{(class}
\keyword{(sA[j]}
\keyword{+}
\keyword{,}
\keyword{=}
\keyword{Add}
\keyword{BinDat}
\keyword{Consider}
\keyword{Could}
\keyword{Data}
\keyword{FALSE)}
\keyword{S3}
\keyword{Second}
\keyword{Should}
\keyword{To}
\keyword{Will}
\keyword{a}
\keyword{able}
\keyword{and}
\keyword{be}
\keyword{case}
\keyword{class}
\keyword{closure}
\keyword{convert}
\keyword{created}
\keyword{created?}
\keyword{data}
\keyword{data.frame}
\keyword{design}
\keyword{dispatch}
\keyword{dispatch,}
\keyword{dmat?}
\keyword{else}
\keyword{faster}
\keyword{field)}
\keyword{for}
\keyword{full}
\keyword{getXmat(),}
\keyword{in}
\keyword{include}
\keyword{internal}
\keyword{is}
\keyword{its}
\keyword{just}
\keyword{long}
\keyword{making}
\keyword{mat}
\keyword{matrix}
\keyword{measure}
\keyword{method}
\keyword{methods}
\keyword{necessary}
\keyword{needed.}
\keyword{not}
\keyword{object}
\keyword{on}
\keyword{one}
\keyword{outcome(s)}
\keyword{passing}
\keyword{perform}
\keyword{possible}
\keyword{previously}
\keyword{reference}
\keyword{return}
\keyword{sW,sA[j]).}
\keyword{save}
\keyword{storage}
\keyword{store}
\keyword{store/create}
\keyword{storing}
\keyword{subset}
\keyword{summary}
\keyword{that}
\keyword{the}
\keyword{this}
\keyword{time}
\keyword{to}
\keyword{transformations}
\keyword{used}
\keyword{various}
\keyword{ways}
\keyword{whatever}
\keyword{when}
\keyword{which}
\keyword{wide}
\keyword{will}

