% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinOutModel_class.R
\docType{class}
\name{BinOutModel}
\alias{BinOutModel}
\title{R6 class for fitting and making predictions for a single logistic regression with binary outcome B, P(B | PredVars)}
\format{An \code{\link{R6Class}} generator object}
\usage{
BinOutModel
}
\description{
This R6 class can request, store and manage the design matrix Xmat, as well as the binary outcome Bin for the
 logistic regression P(Bin|Xmat).
 Can also be used for converting data in wide format to long when requested,
 e.g., when pooling across binary indicators (fitting one pooled logistic regression model for several indicators)
 The class has methods that perform queries to data storage R6 class DatNet.sWsA to get appropriate data columns & row subsets
}
\details{
\itemize{
\item{cont.sVar.flag} - Matrix of friend indices (network IDs) of \code{dim = (nobs x Kmax)}.
\item{bw.j} - Vector of integers, where \code{nF[i]} is the integer number of friends (0 to \code{Kmax}) for observation \code{i}.
\item{glmfitclass} - Number of observations
\item{bindat} - Maximum number of friends for any observation.
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg)}}{Uses \code{reg} R6 \code{\link{RegressionClass}} object to instantiate a new model for a
  logistic regression with binary outcome.}
  \item{\code{show()}}{Print information on outcome and predictor names used in this regression model}
  \item{\code{fit()}}{...}
  \item{\code{copy.fit()}}{...}
  \item{\code{predict()}}{...}
  \item{\code{copy.predict()}}{...}
  \item{\code{predictAeqa()}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{getoutvarnm}}{...}
  \item{\code{getoutvarval}}{...}
  \item{\code{getsubset}}{...}
  \item{\code{getprobA1}}{...}
  \item{\code{getfit}}{...}
  \item{\code{wipe.alldat}}{...}
}
}
\keyword{R6}
\keyword{class}

