% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Define_sVar_class.R
\name{def.sW}
\alias{+.Define_sVar}
\alias{def.sA}
\alias{def.sW}
\title{Define Summary Measures sA and sW}
\usage{
def.sW(...)

def.sA(...)

\method{+}{Define_sVar}(sVar1, sVar2)
}
\arguments{
\item{...}{Named R expressions or character strings that specify the formula for creating the summary measures.}

\item{sVar1}{An object returned by a call to \code{def.sW} or \code{def.sA} functions.}

\item{sVar2}{An object returned by a call to \code{def.sW} or \code{def.sA} functions.}
}
\value{
R6 object of class \code{Define_sVar} which must be passed as argument to \code{\link{tmlenet}}.
}
\description{
Define and store \code{tmlenet} function arguments \code{sW} and \code{sA},
 which are arbitrary summary measures based on treatmet \code{Anode} and baseline covariates in \code{Wnodes}
 constructed from the input \code{data}.
 Each summary measure \code{sVar} is specified as a named R expression or a named character argument.
 Separate calls to \code{def.sW/def.sA} functions can be aggregated into a single collection with '+' function,
 e.g., \code{def.sW(W1=W1)+def.sW(W2=W2)}.
 Two special named arguments can be passed to the \code{def.sW}, \code{def.sA} functions:
 \itemize{
 \item \code{noname = TRUE} - to not use the summary measure name provided by the user when assigning
   the names to the summary measure columns (variables); and
 \item \code{replaceNAw0 = TRUE} - To automatically replace all the missing network covariate values (\code{NA}s) with \code{0}.
}
}
\section{Details}{

The R expressions passed to these functions are evaluated later inside \code{\link{tmlenet}} or \code{\link{eval.summaries}} functions,
using the environment of the input data frame and enclosed within the user-calling environment.
}
\examples{
#***************************************************************************************
# Examples of summary measures defined by def.sW
#***************************************************************************************
# Use R expressions to define summary measure functions, s.a.,
# W2[[0]] just means W2; W2[[1:Kmax]] means vectors of (W2_netF_j), j=1, ..., Kmax;
def_sW <- def.sW(netW1 = W1[[0:Kmax]], noname = TRUE) +
          def.sW(netW2 = W2[[1:Kmax]], noname = TRUE) +
          def.sW(netW3 = W3[[1:Kmax]], noname = TRUE) +
          def.sW(netW3_sum = rowSums(W3[[1:Kmax]]), replaceNAw0 = TRUE)

def_sW <- def.sW(netW2 = W2[[1:Kmax]]) +
            def.sW(netW3_sum = rowSums(W3[[1:Kmax]]), replaceNAw0 = TRUE)

def_sW <- def.sW(netW2 = W2[[1:Kmax]], netW3_sum = rowSums(W3[[1:Kmax]]), noname= TRUE,
                  replaceNAw0 = TRUE)
def_sW <- def.sW(netW2 = W2[[1:Kmax]], netW3_sum = rowSums(W3[[1:Kmax]]),
                  replaceNAw0 = FALSE, noname = TRUE)

def_sW <- def.sW(W1 = W1[[0]], W1 = W2[[0:Kmax]], W1 = W3[[0:Kmax]],
                  netW1_sum = rowSums(W1[[1:Kmax]]),
                  netW2_sum = rowSums(W2[[1:Kmax]]),
                  netW3_sum = rowSums(W3[[1:Kmax]]))

#***************************************************************************************
# Examples of summary measures defined by def.sA
#***************************************************************************************
def_sA <- def.sA(sum_1mAW2_nets = rowSums((1-A[[1:Kmax]]) * W2[[1:Kmax]]),
            replaceNAw0 = TRUE) +
            def.sA(netA = A[[0:Kmax]])

#***************************************************************************************
# Example. Three equivalent ways to define summary measures:
#***************************************************************************************
def_sW <- def.sW(W1 = W1[[0]], W2 = W2[[0]], W3 = W3[[0]])
def_sW <- def.sW(W1 = W1, W2 = W2, W3 = W3) # vs 2
def_sW <- def.sW(W1 = "W1[[0]]", W2 = "W2[[0]]", W3 = "W3[[0]]") # vs 3

#***************************************************************************************
# Example. See contents of the Define_sVar object returned by def.sW
#***************************************************************************************
def_sW$sVar.exprs
def_sW$sVar.expr.names
def_sW$ReplMisVal0
def_sW$sVar.misXreplace
def_sW$sVar.noname

#***************************************************************************************
# Example. More summary measures for sA
#***************************************************************************************
def_sA <- def.sA(netA = A[[0:Kmax]], noname = TRUE)
def_sA <- def.sA(netA = "A[[0:Kmax]]",
                sum_1mAW2_nets = rowSums((1-A[[1:Kmax]]) * W2[[1:Kmax]]))

#***************************************************************************************
# Example. More summary measures for sW and sA
#***************************************************************************************
def_sW <- def.sW(netW2 = W2[[1:Kmax]], noname = TRUE) + 
          def.sW(netW3_sum = rowSums(W3[[1:Kmax]]), replaceNAw0 = TRUE)
            
def_sA <- def.sA(sum_1mAW2_nets = rowSums((1-A[[1:Kmax]]) * W2[[1:Kmax]]),
            replaceNAw0 = TRUE) +
          def.sA(netA = A[[0:Kmax]], noname = TRUE)

#***************************************************************************************
# Evaluate the summary measures applied to the  (O)bserved data (data.frame) and network
#***************************************************************************************
data(df_netKmax6) # load observed data
data(NetInd_mat_Kmax6)  # load the network ID matrix
res <- eval.summaries(sW = def_sW, sA = def_sA,  Kmax = 6, data = df_netKmax6,
  NETIDmat = NetInd_mat_Kmax6, verbose = TRUE)

#***************************************************************************************
# Contents of the list returned by eval.summaries():
#***************************************************************************************
names(res)
# matrix of sW summary measures:
res$sW.matrix
# matrix of sA summary measures:
res$sA.matrix
# matrix of network IDs:
res$NETIDmat
# Observed data (sW,sA) stored as "DatNet.sWsA" R6 class object:
res$DatNet.ObsP0
class(res$DatNet.ObsP0)

}
\seealso{
\code{\link{eval.summaries}} for
 evaluation and validation of the summary measures,
 \code{\link{tmlenet}} for estimation.
}

