% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parserfunctions_R6.r
\docType{class}
\name{Define_sVar}
\alias{Define_sVar}
\title{R6 class for parsing and evaluating user-specified summary measures (in \code{sVar.exprs})}
\format{An \code{\link{R6Class}} generator object}
\usage{
Define_sVar
}
\description{
This R6 class can parse and evaluate (given the input data) the summary measures defined with functions
 \code{\link{def.sW}} and \code{\link{def.sA}}.
 The summary measure expressions (stored in \code{sVar.exprs}) are evaluated in the environment of the input data.frame.
 Note that the resulting summary measures are never stored inside this class,
 the data is stored only in \code{\link{DatNet}} and \code{\link{DatNet.sWsA}} R6 classes.
}
\details{
\itemize{
\item{\code{data.df}} - Input data frame that will be used for evaluating the summary measure expressions.
\item{\code{Kmax}} - Maximum number of friends for any observation.
\item{\code{netind_cl}} - Pointer to a network instance of class \code{simcausal::NetIndClass}.
\item{\code{ReplMisVal0}} - A logical vector that captures args \code{replaceNAw0=TRUE/FALSE} in functions \code{def.sW}, \code{def.sA}.
 If \code{TRUE} for a particular summary measure in \code{sVar.exprs} then all missing network \code{VarNode}
 values (when \code{nF[i] < Kmax}) will get replaced with \code{tmlenet:::gvars$misXreplace} (default is 0).
\item{\code{sVar.misXreplace}} - Maximum number of friends for any observation.
\item{\code{sVar.exprs}} - Deparsed summary measure expressions expressions (character vector).
 (one name can represent a multivariate summary measure).
\item{\code{user.env}} - Captured user-environment from which \code{def.sW} or \code{def.sA} was called.
}
}
\section{Methods}{

\describe{
  \item{\code{new(..., type, user.env)}}{...}
  \item{\code{df.names(data.df)}}{List of variables in the input data \code{data.df} gets assigned to a special
  variable (\code{ANCHOR_ALLVARNMS_VECTOR_0})}
}
}
\keyword{R6}
\keyword{class}

