% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummariesModel_class.R
\docType{class}
\name{SummariesModel}
\alias{SummariesModel}
\title{Class for defining, holding and fitting collections of summary measure models P(sA[j] | sW, sA[j])}
\format{An R6 class object.}
\usage{
SummariesModel
}
\description{
Class for defining, holding and fitting collections of summary measure models P(sA[j] | sW, sA[j])
}
\details{
Following fields are created during initialization
\itemize{
\item{n_regs} ...
\item{nodes} ...
\item{sA_nms} ...
\item{sW_nms} ...
}
More details about the class...
}
\keyword{datasets}

