% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmlenet-package.R
\docType{package}
\name{tmlenet-package}
\alias{tmlenet-package}
\title{Targeted Maximum Likelihood Estimation for Networks}
\description{
The \pkg{tmlenet} R package implements the Targeted Maximum Likelihood Estimation (TMLE) of causal effects
 under single time point stochastic interventions in network data.
 The package also implements the Horvitz-Thompson estimator for networks and the parametric g-computation formula-based estimator.
 The inference for the TMLE is based on the efficient influence curves for dependent data.
 For additional details and examples please see the function-specific documentation.
}
\section{Routines}{

The following routines will be generally invoked by a user, in the same order as presented below.
\describe{

\item{\code{\link{def.sW}}}{Defines baseline summary measures \code{sW} that are functions of each unit's baseline covariates
   in \code{Wnodes} as well as unit's friends baseline covariates (\code{Wnode} values of units' friends).
   This is the first part of the two part specification of the structural equation model for the outcome \code{Y}.
   The function \code{def.sW} can take any number of arguments, with each argument specifying either a vector of summary measures for all units
   or a multivariate summary measure in a form of a matrix.
   Each argument to \code{def.sW} must be named and has to be a valid R expression or a string containing an R expression.
   The expressions can contain double bracket indexing to reference variable values of friends. For example,
   \code{Var[[1]]} produces a summary measure that is a vector of length \code{nrow(data)} that contains the values of \code{Var} variable of
   each unit's first friend. One could also use vectors for indexing friends, using a special constant \code{Kmax} to represent the largest number of friends for any unit.
   For example, \code{Var[[1:Kmax]]} produces a summary measure that is a matrix of dimension with nrow(data) rows and Kmax columns,
   where the first column will be equal to \code{Var[[1]]}, and so on, up to the last column being equal to \code{Var[[Kmax]]}. Note that \code{Kmax} can vary depending
   on a particular dataset and is determined by the input network structure. That is \code{Kmax} is the largest node degree observed among all observations.
   By default, \code{NA} is used for the missing network summary measure values, however \code{NA}'s can be automatically replaced with 0's
   by passing \code{replaceNAw0 = TRUE} as an argument to \code{def.sW} function.}
\item{\code{\link{def.sA}}}{Defines treatment summary measures \code{sA} that can be functions of each unit's exposure & baseline covariates,
   as well the exposures and baseline covariates of unit's friends.
   This is the second part of the two part specification of the structural equation model for the outcome \code{Y}.
   The syntax is identical to \code{def.sW} function except that \code{def.sA} can contain any summary measures defined as functions of \code{Wnodes}
   and \code{Anode}.}
\item{\code{\link{eval.summaries}}}{A convenience function that may be used for checking and validating the user-defined summary measures.
   Takes the input dataset and evaluates the summary measures based on objects previously defined with function calls \code{def.sW} and \code{def.sA}.
   Note that this function is automatically executed by \code{tmlenet} and does not need to be called by the user prior to calling \code{tmlenet}.}
\item{\code{\link{tmlenet}}}{Performs estimation of the causal effect of interest using the observed input \code{data},
   the intervention of interest, the network information and the previously defined summary measures \code{sW}, \code{sA}.}
}
}

\section{Datasets}{

To learn more about the type of data input required by \code{\link{tmlenet}}, see the following example datasets:
\itemize{
  \item \code{\link{df_netKmax2}}
  \item \code{\link{df_netKmax6}}
  \item \code{\link{NetInd_mat_Kmax6}}
}
}

\section{Updates}{

Check for updates and report bugs at \url{http://github.com/osofr/tmlenet}.
}

